==> analysis/0-preprocessing.Rmd <==
---
title: "Preprocessing"
author: "Helena L. Crowell"
date: "`r format(Sys.Date(), '%B %d, %Y')`"
---

```{r setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(cache = TRUE, autodep = TRUE, cache.lazy = FALSE)
```

# Preamble

## Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(biomaRt)
library(dplyr)
library(scater)
library(scran)
library(SingleCellExperiment)
```

## Load data

```{r load-data, message = FALSE, warning = FALSE}
counts <- readRDS(file.path("data", "UMI_counts_merged.rds"))
```

## Construct SCE

```{r make-sce}
ss <- strsplit(colnames(counts), "_")
pids <- vapply(ss, .subset, i = 1, character(1))
cd <- DataFrame(plate_id = pids, row.names = colnames(counts))

(sce <- SingleCellExperiment(
    assays = list(counts = counts),
    colData = cd))
```

# QC & filtering

## Get gene symbols

```{r get-symbols}
ensembl <- useMart("ensembl", 
    dataset = "drerio_gene_ensembl",
    host = "dec2017.archive.ensembl.org")

map <- getBM(
    attributes = c("ensembl_gene_id", "zfin_id_symbol", "chromosome_name"),
    filter = "ensembl_gene_id", value = rownames(sce), mart = ensembl)

# check that all genes are annotated
setdiff(rownames(sce), map$ensembl_gene_id)

# replace problematic characters
rownames(sce) <- gsub("_", "-", rownames(sce))

# add gene symbols to SCE rownames
m <- match(rownames(sce), map$ensembl_gene_id)
s <- map$zfin_id_symbol[m]
rownames(sce) <- paste(rownames(sce), s, sep = ":")
rowData(sce)$symbol <- s
```

## Get MT genes

```{r get-mito}
mt <- map$chromosome_name == "MT"
mt <- rownames(sce)[mt[m]]
(mt <- mt[!is.na(mt)])
# set MT genes as controls
altExp(sce, "mt") <- sce[mt, ]
```

## Compute QC metrics

```{r calc-qc, fig.width = 6, fig.height = 4}
# drop undetected genes
sce <- sce[rowSums(counts(sce) > 0) > 0, ]

# calculate feature- & cell-level QC metrics
sce <- addPerFeatureQC(sce)
sce <- addPerCellQC(sce)
```

## Filter outliers

```{r fil-ols, fig.width = 5, fig.height = 3}
table(sce$ol <- isOutlier(sce$detected, 
    nmads = 2, type = "both", log = TRUE))

as.data.frame(colData(sce)) %>% 
    ggplot(aes(x = detected, alpha = ol)) + 
        scale_alpha_manual(values = c("TRUE" = 0.2, "FALSE" = 0.8)) +
        geom_histogram(bins = 100) + theme_bw()

sce <- sce[, !sce$ol]
table(sce$plate_id)
```

## QC plots

### Total features vs. %(MT)

```{r fig.width = 5, fig.height = 3}
fix_lgd <- guides(fill = guide_legend(
  override.aes = list(size = 3, alpha = 1)))
plotColData(sce, 
    x = "detected", y = "altexps_mt_percent", 
    colour_by = "plate_id", point_alpha = 0.8) + 
    scale_x_log10() + fix_lgd
```

### Distributions of %(MT) by plate

```{r fig.width = 5, fig.height = 3}
plotColData(sce, x = "plate_id", y = "altexps_mt_percent", 
    colour_by = "plate_id", show_median  = TRUE) + fix_lgd
```

### Highest expressing features

```{r fig.width = 6, fig.height = 5}
plotHighestExprs(sce, colour_cells_by = "detected")
```

## Normalization

```{r}
# compute log-normalized counts using MT genes as controls
sce <- sce[rowSums(counts(sce) > 1) >= 10, ]
sce <- computeSumFactors(sce)
sce <- computeSpikeFactors(sce, "mt")
sce <- logNormCounts(sce)
```

### Expression frequency vs. mean

```{r message = FALSE, fig.width = 5, fig.height = 3}
rowData(sce)$mean_expr <- rowMeans(logcounts(sce))
rowData(sce)$is_mt <- rownames(sce) %in% mt
plotRowData(sce, x = "mean_expr", y = "detected", colour_by = "is_mt") +
    geom_hline(yintercept = 50, lty = 2) +
    geom_text(x = 5, y = 45, label = paste(
        sum(rowData(sce)$detected >= 50),
        "genes expressed ≥ 50% of cells")) +
    geom_hline(yintercept = 25, lty = 3) +
    geom_text(x = 5, y = 20, label = paste(
        sum(rowData(sce)$detected >= 25),
        "genes expressed ≥ 25% of cells")) + fix_lgd
```

# Dimension reduction

```{r}
# run PCA & t-SNE
sce <- runPCA(sce) 
sce <- runTSNE(sce)
```

## PCA colored by plate ID & library size

```{r pca, fig.width = 10, fig.height = 4}
sce$log10_total <- log10(sce$total + 1)
multiplot(cols = 2,
    plotPCA(sce, colour_by = "plate_id") + fix_lgd,
    plotPCA(sce, colour_by = "log10_total"))
```

## t-SNE colored by plate ID & %(MT)

```{r tsne, fig.width = 10, fig.height = 4}
multiplot(cols = 2,
    plotTSNE(sce, colour_by = "plate_id") + fix_lgd,
    plotTSNE(sce, colour_by = "altexps_mt_percent"))
```

# Save data

## mCherry expression

```{r fig.width = 5, fig.height = 3}
plotExpression(sce, 
    features = "drl-mCherry:NA", x = "plate_id", 
    colour_by = "plate_id", show_median = TRUE) + fix_lgd
```

## Write SCE to .rds

```{r}
# exclude low-quality plate 4
sce <- sce[, sce$plate_id != "SN4"]
saveRDS(sce, file.path("output", "0-sce_prep.rds"))
```
