==> analysis/2-annotation.Rmd <==
---
title: "Annotation"
author: "Helena L. Crowell"
date: "`r format(Sys.Date(), '%B %d, %Y')`"
---

```{r setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(cache = TRUE, autodep = TRUE, cache.lazy = FALSE)
```

# Preamble

## Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(circlize)
library(ComplexHeatmap)
library(cowplot)
library(dplyr)
library(ggplot2)
library(readxl)
library(reshape2)
library(scater)
library(scran)
library(Seurat)
library(SingleCellExperiment)
library(tidyr)
```

## Load data

```{r load-data}
sce <- readRDS(file.path("output", "1-sce_clust.rds"))
so <- readRDS(file.path("output", "1-so_clust.rds"))

source(file.path("code", "utils.R"))
mgs_canonical 

m <- match(mgs_canonical, rowData(sce)$symbol, nomatch = 0)
gs <- rownames(sce)[m]
```

## Annotation

```{r anno}
# load cluster annotations
anno <- read_xlsx(file.path("data", "cluster_anno.xlsx"))
anno <- setNames(anno$label, anno$cluster_id)
anno <- anno[as.character(sce$cluster_id)]
# set cluster IDs to manual annotation assignments
Idents(so) <- sce$cluster_id <- droplevels(
    factor(anno, levels = names(cluster_pal)))
```

## Prep. data

```{r}
# scale expression between 0 and 1 using
# (1%)- and (99%)-quantiles as boundaries
es <- .scale(logcounts(sce))

# prep. data.frame of reduced dimensions, 
# cell metadata, and expression for plotting
dr <- do.call(cbind, reducedDims(sce))
df <- data.frame(t(es), dr, colData(sce), 
    i = seq_len(ncol(sce)), check.names = FALSE)
# randomize cell ordering
df <- df[sample(nrow(df), nrow(df)), ] 
```

# Dim. red.

```{r umap-cluster_id, fig.width = 5, fig.height = 3}
# plot UMAP colored by cluster ID
(p <- ggplot(df, aes(x = UMAP_1, y = UMAP_2, col = cluster_id)) +
    geom_point(size = 0.8, alpha = 0.8) + 
    scale_color_manual(NULL, values = cluster_pal) + 
    guides(color = guide_legend(override.aes = list(alpha = 1, size = 2))) +
    theme_void() + theme(aspect.ratio = 1))
```

```{r save-umap-cluster_id, echo = FALSE}
p <- p + theme_void(base_size = 8) + theme(legend.key.height = unit(2, "mm"))
ggsave(file.path("figures", "umap-cluster_id.pdf"), p,
    width = 12, height = 8, units = "cm", useDingbats = FALSE)
```

# Canonical markers

## Dim. red.

```{r umap-mgs, fig.width = 12, fig.height = 14}
ps <- lapply(gs, function(g) { df$expr <- df[, g]
    ggplot(df, aes(x = UMAP_1, y = UMAP_2, col = expr)) +
        geom_point(size = 0.8, alpha = 0.8) + 
        scale_color_viridis_c(NULL, option = "C") +
        ggtitle(g) + theme_void(base_size = 6) + 
        theme(aspect.ratio = 1, legend.key.width = unit(2, "mm"))
})
plot_grid(plotlist = ps, ncol = 5)
```

```{r save-umap-mgs, echo = FALSE, results = "hide"}
p <- ps[[1]] +  theme(legend.key.height = unit(3, "mm")) +
    scale_color_viridis_c("scaled\nexpression", option = "C", limits = c(0,1))
lgd <- get_legend(p)
ps <- lapply(ps, function(p) {
    p$layers[[1]]$aes_params$size <- 0
    p + theme(plot.title = element_text(size = 5),
        legend.position = "none")
})
pdf(file.path("figures", "umap-markers.pdf"), 
    width = 15/2.54, height = 18/2.54)
plot_grid(plotlist = ps, lgd, ncol = 5)
dev.off()
```

## Dot plot

```{r dotplot-mgs, warning = FALSE, message = FALSE, fig.width = 10, fig.height = 4}
(p <- .dotplot(so, gs))
```

```{r save-dotplot-mgs, echo = FALSE, message = FALSE}
p <- p + scale_size_continuous(range = c(0, 3)) + 
    theme_minimal(base_size = 8) + theme(
        legend.key.size = unit(3, "mm"),
        panel.grid = element_blank(),
        axis.title = element_blank(),
        axis.text = element_text(color = "black"),
        axis.text.x = element_text(angle = 45, hjust = 1))
ggsave(file.path("figures", "dotplot-canonical_markers.pdf"),
    width = 15, height = 6, units = "cm", useDingbats = FALSE)
```

# Other markers

```{r umap-other, fig.width = 6, fig.height = 4}
fs <- c("meis2b", "meis3", "foxh1", "sfrp5", "gata5", "gata6")
m <- match(fs, rowData(sce)$symbol)
ps <- lapply(rownames(sce)[m], function(g) { 
    df$expr <- df[, g]
    ggplot(df, aes(x = UMAP_1, y = UMAP_2, col = expr)) +
        geom_point(size = 0, alpha = 0.8) + 
        scale_color_viridis_c(NULL, option = "C") +
        ggtitle(g) + theme_void(base_size = 6) + 
        theme(aspect.ratio = 1, 
            legend.position = "none",
            plot.title = element_text(size = 5))
})
(p <- plot_grid(plotlist = ps, ncol = 3))
```

```{r save-umap-other, echo = FALSE, results = "hide"}
pdf(file.path("figures", "umap-more_genes.pdf"), 
    width = 8/2.54, height = 6/2.54)
p; dev.off()
```

# hox genes

```{r get-hox-genes}
is_hox <- grepl("hox", rowData(sce)$symbol)
hox_gs <- rownames(sce)[is_hox]
```

## Dot plot

```{r dotplot-hox, warning = FALSE, message = FALSE, fig.width = 8, fig.height = 5}
(p <- .dotplot(so, hox_gs))
```

```{r save-dotplot-hox, echo = FALSE, warning = FALSE, message = FALSE}
p <- p + scale_size_continuous(range = c(0, 3)) + 
    theme_minimal(base_size = 8) + theme(
        legend.key.size = unit(3, "mm"),
        panel.grid = element_blank(),
        axis.title = element_blank(),
        axis.text = element_text(color = "black"),
        axis.text.x = element_text(angle = 45, hjust = 1))
ggsave(file.path("figures", "dotplot-hox_genes.pdf"),
    width = 12, height = 6, units = "cm", useDingbats = FALSE)
```

## Dim. red.

```{r umap-hox, fig.width = 16, fig.height = 7}
ps <- lapply(hox_gs, function(g) { df$expr <- df[, g]
    ggplot(df, aes(x = UMAP_1, y = UMAP_2, col = expr)) +
        geom_point(size = 0.8, alpha = 0.8) + 
        scale_color_viridis_c(NULL, option = "C") +
        ggtitle(g) + theme_void(base_size = 6) + 
        theme(aspect.ratio = 1, legend.key.width = unit(2, "mm"))
})
plot_grid(plotlist = ps, ncol = 7)
```

```{r save-umap-hox, echo = FALSE, results = "hide"}
ps <- lapply(ps, function(p) {
    p$layers[[1]]$aes_params$size <- 0
    p + theme(plot.title = element_text(size = 5),
        legend.position = "none")
})
pdf(file.path("figures", "umap-hox_genes.pdf"), 
    width = 20/2.54, height = 8/2.54)
plot_grid(plotlist = ps, ncol = 7)
dev.off()
```

# hand2+ markers

## Merging

```{r merging}
# merge into super-clusters
old <- levels(sce$cluster_id)
ss <- strsplit(old, "_|-")
new <- sapply(ss, .subset, 1)
m <- match(sce$cluster_id, old)
sce$merging_id <- factor(new[m], unique(new))

# exclude endoderm cluster
sub <- sce[, sce$merging_id != "endoderm"]
sub$merging_id <- droplevels(sub$merging_id)

# construct color palette for merged clusters
i <- sapply(levels(sub$merging_id), grep, names(cluster_pal))
i <- sapply(i, function(u) u[ceiling(length(u)/2)])
merging_pal <- setNames(cluster_pal[i], levels(sub$merging_id))
```

## DE analysis

```{r de-analysis}
# run DE analysis
res <- findMarkers(sub, sub$merging_id, pval.type = "some")
df <- data.frame(res$hand2)

# filter at FDR 5%, average logFC > 0
df <- df[df$FDR < 0.05, , drop = FALSE]
df <- df[df$summary.logFC > 0, , drop = FALSE]
df <- top_n(df, 50, summary.logFC)
df$gene <- rownames(df)
m <- match(df$gene, rownames(sce))
df$symbol <- rowData(sce)$symbol[m]
length(mgs <- df$gene)
```

## Heatmap

```{r heatmap-mgs, fig.width = 8, fig.height = 3}
# compute mean expression by cluster
names(mids) <- mids <- levels(sub$merging_id)
cs <- split(seq_len(ncol(sub)), sub$merging_id)
ms <- lapply(mids, function(m) 
    rowMeans(logcounts(sub[mgs, cs[[m]]])))

# use gene symbols as names
mat <- scale(t(do.call(cbind, ms)))
m <- match(mgs, df$gene)
colnames(mat) <- ifelse(
    df$symbol[m] == "" | is.na(df$symbol[m]), 
    df$gene[m], df$symbol[m])

# construct row annotation
row_anno <- rowAnnotation(
    simple_anno_size = unit(3.5, "mm"),
    df = data.frame(merging_id = mids),
    col = list(merging_id = merging_pal),
    show_legend = FALSE,
    gp = gpar(col = "white"),
    show_annotation_name = FALSE,
    annotation_legend_param = list(
        title_gp = gpar(fontsize = 8),
        labels_gp = gpar(fontsize = 6)))

# specify heatmap color function
rng <- max(abs(range(mat, na.rm = TRUE)))
rng <- ceiling(rng*0.5)/0.5
hm_pal <- c("navy", "royalblue", "grey95", "yellowgreen", "darkgreen")
hm_pal <- colorRamp2(seq(-rng, rng, l = 5), hm_pal)

(hm <- Heatmap(mat,
    col = hm_pal,
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    rect_gp = gpar(col = "white"),
    heatmap_legend_param = list(
        title = "scaled mean\nexpression",
        title_gp = gpar(fontsize = 8),
        labels_gp = gpar(fontsize = 6)),
    column_split = df$cluster,
    row_title = NULL,
    column_title = NULL,
    left_annotation = row_anno,
    row_names_gp = gpar(fontsize = 8),
    column_names_gp = gpar(fontsize = 8)))
```

```{r save-heatmap-mgs, echo = FALSE}
hm <- grid.grabExpr(draw(hm))
fn <- file.path("figures", "heatmap-hand2_mgs.pdf")
ggsave(fn, hm, width = 24, height = 6, units = "cm")
```

# Save data

```{r save-data}
saveRDS(so, file.path("output", "2-so_anno.rds"))
saveRDS(sce, file.path("output", "2-sce_anno.rds"))
```
